Move isolated tag-specific views to tags module
